[ET-11] FIX: Add image to Product constructor
We have added the IMAGE column to the PRODUCT table as part of the
ProductTile enhancement. The column was specified as non-null, but the
test for /api/v1/products did not end up initializing the image field of
the entity it wanted to insert, causing a database integrity exception.

Going forward, every pull-request will have Snap CI build its branch to
check test status before merging into master.